refactor: move ResourceStore/SecretManager traits to common (#127)
The postgres storage backend depended on the server crate purely because
the ResourceStore/ResourceStoreReader/SecretManager traits it implements
lived in server — an inverted dependency that pulled axum/HTTP-server logic
into a backend.

Move those pure abstractions down into common (behind a new off-by-default
`store` feature) so backends depend on the abstraction, not on the server:

- common: add store.rs (ResourceStore, ResourceStoreReader,
  ProvidesResourceStore, ProvidesObjectStore, ObjectStoreAdapter) and
  services/secrets.rs (SecretManager, ProvidesSecretManager); add a bare
  AlreadyExists error variant wired into error_code/response_parts.
- server: store.rs and services/secrets.rs become re-export shims of the
  common traits, so existing paths keep working; memory.rs and two api call
  sites switch to common's Error/Result.
- postgres: implement the traits from common, retarget the error conversion
  to common::Error, and drop the unitycatalog-server dependency entirely.

Also remove a long-broken postgres integration test that imported the
nonexistent delta_sharing_common crate, along with its now-unused dev
dependency — this was the only thing keeping axum in postgres's tree.

Result: `cargo tree -p unitycatalog-postgres` no longer contains
unitycatalog-server or axum. No behavior change.

Larger crate-boundary cleanups are tracked as follow-ups: #137 (domain /
transport split in server), #138 (extract proxy handlers crate), #139
(split credential_vending per cloud).
